LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/Create-protocol-with-space.any-expected.txt

CONSOLE MESSAGE: Wrong protocol for WebSocket 'ec ho'

PASS Create WebSocket - Pass a valid URL and a protocol string with a space in it - SYNTAX_ERR is thrown
